Move messaging components to module
