+++ examples/eth-dm/src/App.tsx
@@ -2,21 +2,19 @@ import '@ethersproject/shims';
 
 import React, { useEffect, useState } from 'react';
 import './App.css';
-import { Waku, WakuMessage } from 'js-waku';
+import { Waku } from 'js-waku';
 import { ethers } from 'ethers';
 import { Web3Provider } from '@ethersproject/providers';
-import { createPublicKeyMessage, KeyPair } from './crypto';
-import { encode, PublicKeyMessage } from './messages';
-import Messages, { Message } from './Messages';
+import { KeyPair } from './crypto';
+import Messages, { Message } from './messaging/Messages';
 import 'fontsource-roboto';
 import {
   AppBar,
-  Button,
   IconButton,
   Toolbar,
   Typography,
 } from '@material-ui/core';
-import SendMessage from './SendMessage';
+import SendMessage from './messaging/SendMessage';
 import KeyPairHandling from './key_pair_handling/KeyPairHandling';
 import InitWaku from './InitWaku';
 import {
@@ -26,9 +24,7 @@ import {
 } from '@material-ui/core/styles';
 import { teal, purple, green } from '@material-ui/core/colors';
 import WifiIcon from '@material-ui/icons/Wifi';
-
-export const PublicKeyContentTopic = '/eth-dm/1/public-key/json';
-export const DirectMessageContentTopic = '/eth-dm/1/direct-message/json';
+import BroadcastPublicKey from './BroadcastPublicKey';
 
 declare let window: any;
 
@@ -68,7 +64,6 @@ function App() {
   const [waku, setWaku] = useState<Waku>();
   const [provider, setProvider] = useState<Web3Provider>();
   const [ethDmKeyPair, setEthDmKeyPair] = useState<KeyPair | undefined>();
-  const [publicKeyMsg, setPublicKeyMsg] = useState<PublicKeyMessage>();
   const [publicKeys, setPublicKeys] = useState<Map<string, string>>(new Map());
   const [messages, setMessages] = useState<Message[]>([]);
   const [address, setAddress] = useState<string>();
@@ -92,31 +87,6 @@ function App() {
       .getAddress()
       .then((address) => setAddress(address));
   });
-
-  const broadcastPublicKey = () => {
-    if (!ethDmKeyPair) return;
-    if (!provider) return;
-    if (!waku) return;
-
-    if (publicKeyMsg) {
-      const wakuMsg = encodePublicKeyWakuMessage(publicKeyMsg);
-      waku.lightPush.push(wakuMsg).catch((e) => {
-        console.error('Failed to send Public Key Message', e);
-      });
-    } else {
-      createPublicKeyMessage(provider.getSigner(), ethDmKeyPair.publicKey)
-        .then((msg) => {
-          setPublicKeyMsg(msg);
-          const wakuMsg = encodePublicKeyWakuMessage(msg);
-          waku.lightPush.push(wakuMsg).catch((e) => {
-            console.error('Failed to send Public Key Message', e);
-          });
-        })
-        .catch((e) => {
-          console.error('Failed to creat Eth-Dm Publication message', e);
-        });
-    }
-  };
 
   return (
     <ThemeProvider theme={theme}>
@@ -153,14 +123,11 @@ function App() {
                 ethDmKeyPair={ethDmKeyPair}
                 setEthDmKeyPair={(keyPair) => setEthDmKeyPair(keyPair)}
               />
-              <Button
-                variant="contained"
-                color="primary"
-                onClick={broadcastPublicKey}
-                disabled={!ethDmKeyPair || !waku}
-              >
-                Broadcast Eth-DM Public Key
-              </Button>
+              <BroadcastPublicKey
+                signer={provider?.getSigner()}
+                ethDmKeyPair={ethDmKeyPair}
+                waku={waku}
+              />
             </fieldset>
             <fieldset>
               <legend>Messaging</legend>
@@ -175,8 +142,3 @@ function App() {
 }
 
 export default App;
-
-function encodePublicKeyWakuMessage(ethDmMsg: PublicKeyMessage): WakuMessage {
-  const payload = encode(ethDmMsg);
-  return WakuMessage.fromBytes(payload, PublicKeyContentTopic);
-}
