+++ examples/eth-dm/src/messaging/Messages.tsx
@@ -1,3 +1,6 @@
+/**
+ * Clear text message
+ */
 export interface Message {
   text: string;
   timestamp: Date;
