+++ examples/eth-dm/src/messaging/wire.ts
